useFetch hook implemented
